Refactor MQTT Receiver into Mqtt class directly
