+++ device/transport/mqtt/package.json
@@ -34,7 +34,7 @@
     "alltest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test*.js",
     "ci": "npm -s run lint && npm -s run build && npm -s run alltest-min && npm -s run check-cover",
     "test": "npm -s run lint && npm -s run build && npm -s run unittest",
-    "check-cover": "istanbul check-coverage --statements 94 --branches 80 --functions 88 --lines 95"
+    "check-cover": "istanbul check-coverage --statements 94 --branches 81 --functions 89 --lines 96"
   },
   "engines": {
     "node": ">= 0.10"
